requirements.txt
requests>=2.31.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Flask>=2.3.0
plotly>=5.15.0
nltk>=3.8.0